+++ packages/xchain-litecoin/src/index.ts
@@ -1,4 +1,4 @@
 export * from './types'
 export * from './client'
-export { broadcastTx, getDerivePath, getDefaultFees, getDefaultFeesWithRates, getPrefix } from './utils'
+export { broadcastTx, getDerivePath, getDefaultFees, getDefaultFeesWithRates, getPrefix, LTC_DECIMAL } from './utils'
 export { createTxInfo } from './ledger'