+++ packages/xchain-litecoin/CHANGELOG.md
@@ -5,6 +5,14 @@
 - Update @xchainjs/xchain-client package to 0.4.0
 - Update @xchainjs/xchain-crypto package to 0.2.3
 
+### Update
+
+- Define / Export `LTC_DECIMAL`
+
+### Fix
+
+- Fix derivation path
+
 # v.0.0.2 (2021-01-30)
 
 - Clear lib folder on build